Use a URLFlag and URL.ResolveReference.
This improves type safety. If supplied an invalid URL that fails to url.Parse, then flag.Parse() will provide an error.

Previously, string concatenation was used to construct the final URL, taking "https://sourcegraph.com" + "/-/refs?" + query. But that means if there was a trailing slash in the url, like -sourcegraph_url="https://sourcegraph.com/" then there'd be two slashes in the final url.

Using URL.ResolveReference resolves that.

The only disadvantage is slight additional verbosity of code (but hopefully readability is improved).